fix first time crash

+++ src/config.rs
@@ -364,15 +364,18 @@ impl Opts {
 }
 
 pub fn setup_logger(logfile: &str, level: u8) -> crate::types::Result<()> {
-    let mut suffix = 1;
-    loop {
-        let new_file = logfile.to_string() + "." + suffix.to_string().as_str();
-        let path = Path::new(new_file.as_str());
-        if !path.exists() {
-            std::fs::rename(logfile, new_file.as_str())?;
-            break;
-        } else {
-            suffix += 1;
+    let path = Path::new(logfile);
+    if path.exists() {
+        let mut suffix = 1;
+        loop {
+            let new_file = logfile.to_string() + "." + suffix.to_string().as_str();
+            let path = Path::new(new_file.as_str());
+            if !path.exists() {
+                std::fs::rename(logfile, new_file.as_str())?;
+                break;
+            } else {
+                suffix += 1;
+            }
         }
     }
     let level = match level {
